backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
google-genai>=1.16.0
pillow
python-dotenv
pydantic
numpy
httpx

